feat: buffer-local var for lang override
Adds functionality to override the linting language used by clj-kondo.

When connected to sibling cljs and clj repls in cider, it will use the
major mode to decide where to execute evaluated sexps. When working with
cljc files it can be useful to set the major mode of emacs to
clojure-mode or clojurescript-mode to restrict where the sexps are being
evaluated while you develop.

This PR adds the ability for a buffer to specify the language it would
like to be linted as, allowing cljc buffers being edited in clojure-mode
to still be linted by the cljc linter in clj-kondo.